Strike-through dead link
The link that is struck through is dead.

I was reading the documentation before trying to use Literate.jl and saw this dead link. I'm not sure how to replace it so I just struck it out and submitted this PR, hoping to make it easy for you to spot and edit yourself if you want! 🙂