RFC: Add support for Jupyter cell metadata via %% syntax (#43)
Admittedly this is a little strange: it adds support for a special first-line
syntax that only applies to Jupyter notebook outputs. But this %% format is
somewhat standard across notebook-generating tools, and the ability to protect
it with a `#nb` leader makes it possible to specifically target this line to a
notebook output. We aim to use something like this for generation of notebooks
with support for nbgrader's metadata extensions, but it could also be used to
specify cell types for Jupyter notebook presentations \(with presentation
extensions\) or other such fun. Fixes #67.

Co-authored-by: Matt Bauman <mbauman@gmail.com>
Co-authored-by: Fredrik Ekre <ekrefredrik@gmail.com>