Fixed Local Default
The default path created for the meta block at the top of the literate markdown files did not compile without error when using Documenter. Instead of having an `<unknown>` tag placed, the modification returns pwd() which is consistent with the source when running locally and allows for offline document compilation.