Fix time regression introduced by automatic head branch detection
Automatic head branch detection (introduced in version 2.11.0) caused a
performance regression since the `git remote show` command takes ~1
second. For documentation builds with many literate files this caused
significant slowdowns, which is particularly annoying when doing
iterative buils with eg.
[LiveServer.jl](https://github.com/tlienart/LiveServer.jl). Literate now
caches the remote head branch on a per-repo basis, so the 1 second delay
should only be noticed on the first run of the first file in a repo. As
noted in the changelog entry for 2.11.0 it is also possible to specify
the head branch by passing the `edit_commit` keyword argument. Doing so
will now completely skip the slow `git` command.